+++ lib/hyperglot/main.py
@@ -5,7 +5,7 @@ import yaml
 import logging
 from collections import OrderedDict
 from fontTools.ttLib import TTFont
-from . import __version__, DB, SUPPORTLEVELS
+from . import __version__, DB, SUPPORTLEVELS, VALIDITY
 from .languages import Languages, Language
 
 
@@ -58,48 +58,29 @@ def language_list(langs, native=False, users=False, script=None, strict_iso=Fals
     return seperator.join(items)
 
 
-def print_to_cli(font, title, autonyms, users, script, strict_iso):
+def print_to_cli(font, title, autonyms, users, strict_iso):
     print()
     print("=" * len(title))
     print(title)
     print("=" * len(title))
     print()
-    if "done" in font:
-        done = font["done"]
-        total = 0
-        for script in done:
-            count = len(done[script])
-            if count > 0:
-                print()
-                title = "%d %s of %s script:" % \
-                    (count, "language" if count == 1 else "languages",
-                     script)
-                print(title)
-                print("-" * len(title))
-                print(language_list(done[script],
-                                    autonyms, users, script, strict_iso))
-                total = total + count
-        if total > 0:
+    total = 0
+    for script in font:
+        count = len(font[script])
+        if count > 0:
             print()
-            print("%d languages supported in total." % total)
-            print()
-
-    if "weak" in font:
+            title = "%d %s of %s script:" % \
+                (count, "language" if count == 1 else "languages",
+                    script)
+            print(title)
+            print("-" * len(title))
+            print(language_list(font[script],
+                                autonyms, users, script, strict_iso))
+            total = total + count
+    if total > 0:
         print()
-        weak = font["weak"]
-        for script in weak:
-            count = len(weak[script])
-            if count > 0:
-                print()
-                title = "There are %d %s of %s script that are likely " \
-                    "supported, but the database lacks independent " \
-                    "confirmation:" % \
-                    (count, "langauage" if count == 1 else "languages",
-                     script)
-                print(title)
-                print("-" * len(title))
-                print(language_list(weak[script],
-                                    autonyms, users, script, strict_iso))
+        print("%d languages supported in total." % total)
+        print()
 
 
 def prune_intersect(intersection, res, level):
@@ -173,31 +154,33 @@ MODES = ["individual", "union", "intersection"]
 @click.option("-s", "--support",
               type=click.Choice(SUPPORTLEVELS.keys(), case_sensitive=False),
               default="base", show_default=True,
-              help="What level of language support to check the fonts for."
-              )
+              help="Option to test only for the language's base charset, or to"
+              " also test for presence of all auxilliary characters, if "
+              "present in the database.")
+@click.option("--validity", type=click.Choice(VALIDITY, case_sensitive=False),
+              default=VALIDITY[1], show_default=True,
+              help="The level of validity for languages matched against the "
+              "font. Weaker levels always include more strict levels. The "
+              "default includes all languages for which the database has "
+              "charset data.")
 @click.option("-a", "--autonyms", is_flag=True, default=False,
               help="Flag to render languages names in their native name.")
 @click.option("-u", "--users", is_flag=True, default=False,
               help="Flag to show how many users each languages has.")
 @click.option("-o", "--output", type=click.File(mode="w", encoding="utf-8"),
-              help="Provide a name to a yaml file to write support "
+              help="Provide a name for a yaml file to write support "
               "information to.")
 @click.option("-m", "--mode", type=click.Choice(MODES, case_sensitive=False),
               default=MODES[0], show_default=True,
-              help="When checking more than one file, a comparison can be "
+              help="When passing in more than one file, a comparison can be "
               "generated. By default each file's support is listed "
               "individually. 'union' shows support for all languages "
               "supported by the combination of the passed in fonts. "
-              "'intersection' shows the support all files have in common.")
+              "'intersection' shows the support all fonts have in common.")
 @click.option("--include-historical", is_flag=True, default=False,
               help="Flag to include otherwise ignored historical languages.")
 @click.option("--include-constructed", is_flag=True, default=False,
               help="Flag to include otherwise ignored contructed languages.")
-@click.option("--strict-support", is_flag=True, default=False,
-              help="Flag to exclude language support data that has not "
-              "undergone confirmation through several independend expert "
-              "resources. All language data is carefully compiled from "
-              "reliable sources and can be considered fairly reliable.")
 @click.option("--strict-iso", is_flag=True, default=False,
               help="Flag to display names and macrolanguage data "
               "strictly abiding to ISO data. Without it apply some gentle "
@@ -205,8 +188,8 @@ MODES = ["individual", "union", "intersection"]
               "macrolanguage structure that deviates from ISO data.")
 @click.option("-v", "--verbose", is_flag=True, default=False)
 @click.option("-V", "--version", is_flag=True, default=False)
-def cli(fonts, support, autonyms, users, output, mode, include_historical,
-        include_constructed, strict_support, strict_iso, verbose, version):
+def cli(fonts, support, validity, autonyms, users, output, mode,
+        include_historical, include_constructed, strict_iso, verbose, version):
     """
     Pass in one or more fonts to check their languages support
     """
@@ -220,8 +203,7 @@ def cli(fonts, support, autonyms, users, output, mode, include_historical,
         print("Provide at least one path to a font or --help for more "
               "information")
 
-    # A dict with each file and its results for done and weak, e.g.
-    # { 'filea.otf': { 'done': {..}, 'weak': {..} }, 'fileb.otf: .. }
+    # A dict with each file and its results for each script
     results = {}
 
     for font in fonts:
@@ -230,49 +212,12 @@ def cli(fonts, support, autonyms, users, output, mode, include_historical,
         chars = [chr(c) for c in cmap.getBestCmap().keys()]
         Lang = Languages(strict_iso)
         langs = Lang.get_support_from_chars(
-            chars, include_historical, include_constructed)
-        done = {}
-        weak = {}
-        done_statuses = ["done", "strong"]  # plus "status" not in dict
+            chars, validity, include_historical, include_constructed)
         level = SUPPORTLEVELS[support]
 
-        # Sort the results for this font by db status
-        for script in langs:
-            if script not in done:
-                done[script] = {}
-            if script not in weak:
-                weak[script] = {}
-
-            if level not in langs[script]:
-                continue
-            for iso, l in langs[script][level].items():
-                if "todo_status" not in l or l["todo_status"] in done_statuses:
-                    done[script][iso] = l
-                else:
-                    weak[script][iso] = l
-
         results[font] = {}
+        results[font] = langs
 
-        if strict_support:
-            if done:
-                results[font]["done"] = done
-
-            if weak:
-                results[font]["weak"] = weak
-        else:
-            merged = {}
-            if done:
-                merged = done
-            if weak:
-                if merged == {}:
-                    merged = weak
-                else:
-                    for script, data in weak.items():
-                        if script in merged:
-                            merged[script].update(weak[script])
-                        else:
-                            merged[script] = weak[script]
-            results[font]["done"] = merged
         _font.close()
 
     # Mode for comparison of several files
@@ -280,8 +225,7 @@ def cli(fonts, support, autonyms, users, output, mode, include_historical,
         for font in fonts:
             title = "%s has %s support for:" % (os.path.basename(font),
                                                 level.lower())
-            print_to_cli(results[font], title, autonyms,
-                         users, script, strict_iso)
+            print_to_cli(results[font], title, autonyms, users, strict_iso)
         data = results
     elif mode == "union":
         union = {}
@@ -304,7 +248,7 @@ def cli(fonts, support, autonyms, users, output, mode, include_historical,
 
         title = "Fonts %s together have %s support for:" % \
             (", ".join([os.path.basename(f) for f in fonts]), level.lower())
-        print_to_cli(union, title, autonyms, users, script, strict_iso)
+        print_to_cli(union, title, autonyms, users, strict_iso)
         # Wrap in "single file" 'union' top level, which will be removed when
         # writing the data
         data = {"union": union}
@@ -320,7 +264,7 @@ def cli(fonts, support, autonyms, users, output, mode, include_historical,
 
         title = "Fonts %s all have common %s support for:" % \
             (", ".join([os.path.basename(f) for f in fonts]), level.lower())
-        print_to_cli(intersection, title, autonyms, users, script, strict_iso)
+        print_to_cli(intersection, title, autonyms, users, strict_iso)
         # Wrap in "single file" 'intersection' top level, which will be removed
         # when writing the data
         data = {"intersection": intersection}